requirements.txt
python-dateutil
py4j==0.10.9.7
apache-flink-libraries
apache-flink
pyflink
kafka-python
pymongo
psycopg2-binary
beautifulsoup4
selenium
requests
python-dotenv
feedparser
sqlalchemy
psycopg2-binary
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
eventlet==0.35.1
gunicorn==21.2.0
tensorflow
numpy
pickle-mixin